+++ src/app/api/menu-items/route.js
@@ -1,7 +1,11 @@
 import mongoose from 'mongoose';
 import { MenuItem } from '@/models/MenuItem';
+import { isAdmin } from '../auth/[...nextauth]/route';
 
 export async function POST(req) {
+  if (!(await isAdmin())) {
+    return Response.json(true);
+  }
   mongoose.connect(process.env.MONGO_URL);
   const data = await req.json();
   const menuItemDoc = await MenuItem.create(data);
@@ -9,10 +13,13 @@ export async function POST(req) {
 }
 
 export async function PUT(req) {
+  if (!(await isAdmin())) {
+    return Response.json(true);
+  }
   mongoose.connect(process.env.MONGO_URL);
-  const {_id, ...data} = await req.json();
-  await MenuItem.findByIdAndUpdate(_id, data)
-  return Response.json(true)
+  const { _id, ...data } = await req.json();
+  await MenuItem.findByIdAndUpdate(_id, data);
+  return Response.json(true);
 }
 
 export async function GET() {
@@ -21,10 +28,12 @@ export async function GET() {
 }
 
 export async function DELETE(req) {
+  if (!(await isAdmin())) {
+    return Response.json(true);
+  }
   mongoose.connect(process.env.MONGO_URL);
-  const url = new URL(req.url)
-  const _id = url.searchParams.get('_id')
-  await MenuItem.deleteOne({_id})
-  return Response.json(true)
+  const url = new URL(req.url);
+  const _id = url.searchParams.get('_id');
+  await MenuItem.deleteOne({ _id });
+  return Response.json(true);
 }
-
